backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "MedDash FastAPI and LangGraph backend"
readme = "README.md"
authors = [
    { name = "Benedict", email = "benedict@ekohe.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "langgraph>=0.2.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]

[project.scripts]
backend = "backend.__main__:main"

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
testpaths = ["tests"]
